apps/api/requirements.txt
alembic==1.16.1
annotated-doc==0.0.4
annotated-types==0.7.0
apscheduler>=3.10.0
anyio==4.12.1
bcrypt==4.2.1
click==8.3.1
cryptography==44.0.2
email-validator==2.2.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.7.0
python-dotenv==1.1.0
python-jose==3.4.0
redis==7.3.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
qrcode==8.2
pillow==12.1.1
pillow-heif>=1.3.0
python-slugify==8.0.4
python-multipart==0.0.22
httpx>=0.27.0
